Add Config.toml

+++ src/utils.rs
@@ -14,7 +14,7 @@ pub fn now_timestamp_ms() -> i64 {
     return Utc::now().timestamp_millis();
 }
 
-pub fn is_seconds_passed(required_time_sec: i64, timestamp: i64) -> bool {
+pub fn is_seconds_passed(required_time_sec: i32, timestamp: i64) -> bool {
     let now = now_timestamp_ms();
     let elapsed_time = now - timestamp;
     elapsed_time >= (required_time_sec * 1000) as i64
